cabbooking/src/main/java/com/example/cabbooking/service/RidersManager.java
package com.example.cabbooking.service;

import com.example.cabbooking.exception.RiderAlreadyExistsException;
import com.example.cabbooking.exception.RiderNotFoundException;
import com.example.cabbooking.model.Rider;
import lombok.NonNull;
import org.springframework.stereotype.Service;

import java.util.HashMap;
import java.util.Map;

@Service
public class RidersManager {
    Map<String, Rider> riders = new HashMap<>();

    public void createRider(@NonNull final Rider newRider) {
        if (riders.containsKey(newRider.getId())) {
            throw new RiderAlreadyExistsException();
        }

        riders.put(newRider.getId(), newRider);
    }

    public Rider getRider(@NonNull final String riderId) {
        if (!riders.containsKey(riderId)) {
            throw new RiderNotFoundException();
        }
        return riders.get(riderId);
    }
}
